+++ Main/Include/cont.h
@@ -19,7 +19,7 @@ class continent
  public:
   friend class worldmap;
   continent() { }
-  continent(uchar Index) : Index(Index) { }
+  continent(ushort Index) : Index(Index) { }
   void AttachTo(continent*);
   void Add(vector2d);
   void Save(outputfile&) const;
